+++ packages/desktop/CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fixes issue with reconnecting to peers on resume on iOS ([#2424](https://github.com/TryQuiet/quiet/issues/2424))
 * Fixes references to 'invite code' to be 'invite link' in UI ([#2441](https://github.com/TryQuiet/quiet/issues/2441))
 * Fixes issue with image messages not displaying/throwing errors on iOS ([#2526](https://github.com/TryQuiet/quiet/issues/2526))
+* Removing parsing from channel and community naming ([#2302](https://github.com/TryQuiet/quiet/issues/2302))
 
 # Chores